# fix: overlapping text styling duplicate