--- a/test/__testutils__/createRichTextFixtures.ts
+++ b/test/__testutils__/createRichTextFixtures.ts
@@ -4,13 +4,14 @@ import enRichTextJSON from "../__fixtures__/enRichText.json";
 import cnRichTextJSON from "../__fixtures__/cnRichText.json";
 import koRichTextJSON from "../__fixtures__/koRichText.json";
 import emojiRichTextJSON from "../__fixtures__/emojiRichText.json";
+import overlappedRichTextJSON from "../__fixtures__/overlappedRichText.json";
 
 const deepCloneJSON = <T>(json: T): T => {
 	return JSON.parse(JSON.stringify(json));
 };
 
 export const createRichTextFixtures = (): Record<
-	"en" | "cn" | "ko" | "emoji",
+	"en" | "cn" | "ko" | "emoji" | "overlapped",
 	RichTextField
 > => {
 	return {
@@ -18,5 +19,6 @@ export const createRichTextFixtures = (): Record<
 		cn: deepCloneJSON(cnRichTextJSON) as RichTextField,
 		ko: deepCloneJSON(koRichTextJSON) as RichTextField,
 		emoji: deepCloneJSON(emojiRichTextJSON) as RichTextField,
+		overlapped: deepCloneJSON(overlappedRichTextJSON) as RichTextField,
 	};
 };