<!--- Provide a general summary of your changes in the title above -->

## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Chore (a non-breaking change which is related to package maintenance)
- [x] Bug fix (a non-breaking change which fixes an issue)
- [ ] New feature (a non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Description

<!--- Describe your changes in detail -->
Overlapping text styling starting or ending on the same character is not handled correctly, see: https://github.com/prismicio/prismic-client/issues/198

### Starting case:
This:
```json
{
	"type": "paragraph",
	"text": "Lorem ipsum dolor sit amet",
	"spans": [
		{
			"start": 18,
			"end": 21,
			"type": "em"
		},
		{
			"start": 18,
			"end": 26,
			"type": "strong"
		}
	]
}
```
Leads to:
```html
<p>Lorem ipsum dolor <em>sit</em><strong><em>sit</em> amet</strong></p>
```

### Ending case (theoretically not be possible):
This:
 ```json
{
	"type": "paragraph",
	"text": "Lorem ipsum dolor sit amet",
	"spans": [
		{
			"start": 22,
			"end": 26,
			"type": "em"
		},
		{
			"start": 18,
			"end": 26,
			"type": "strong"
		}
	]
}
```
Leads to:
```html
<p>Lorem ipsum dolor sit <em>amet</em><strong>sit <em>amet</em></strong></p>
```

<!--- Why is this change required? What problem does it solve? -->
<!--- If it resolves an open issue, please link to the issue here. For example "Resolves: #137" -->

### Resolution

The first case above is a valid API response that is actually quite easy to achieve: https://tmp-nuxt-154.cdn.prismic.io/api/v2/documents/search?ref=YaTS7BMAAB8Atqxu, the second is theoretically not possible to achieve(?)

The API sorts nodes ascendingly by `start` index, then ascendingly by `end` index (maybe it doesn't take `end` it into account at all(?))
On the other hand, `asTree()` expects nodes to be sorted ascendingly by `start` index then descendingly by `end` index.

This PR fixes that by skipping spans when next span starts on the same index. It also handles the theoretically impossible case which I'm not sure it's relevant to handle(?)

<details>
<summary>Performances (tl;dr; nothing noticeable)</summary>

Before:
```
v1#asTree x 1,047 ops/sec ±2.31% (86 runs sampled)
v2#asTree x 97,375 ops/sec ±3.40% (85 runs sampled)
Fastest is v2#asTree
v1#asText x 1,068,511 ops/sec ±4.03% (82 runs sampled)
v2#asText x 3,619,421 ops/sec ±4.01% (81 runs sampled)
Fastest is v2#asText
v1#serialize x 958 ops/sec ±3.20% (87 runs sampled)
v2#serialize x 38,952 ops/sec ±4.03% (83 runs sampled)
Fastest is v2#serialize
```

`Array.prototype.sort()` approach from 98762d8:
```
v1#asTree x 1,011 ops/sec ±2.96% (81 runs sampled)
v2#asTree x 96,990 ops/sec ±1.46% (91 runs sampled)
Fastest is v2#asTree
v1#asText x 1,105,438 ops/sec ±3.62% (88 runs sampled)
v2#asText x 3,651,422 ops/sec ±3.02% (85 runs sampled)
Fastest is v2#asText
v1#serialize x 934 ops/sec ±2.37% (85 runs sampled)
v2#serialize x 39,667 ops/sec ±2.29% (89 runs sampled)
Fastest is v2#serialize
```

`continue` (preferred) approach:
```
v1#asTree x 951 ops/sec ±3.55% (81 runs sampled)
v2#asTree x 100,707 ops/sec ±0.97% (91 runs sampled)
Fastest is v2#asTree
v1#asText x 1,158,102 ops/sec ±1.14% (91 runs sampled)
v2#asText x 3,736,161 ops/sec ±1.16% (90 runs sampled)
Fastest is v2#asText
v1#serialize x 944 ops/sec ±1.42% (89 runs sampled)
v2#serialize x 40,375 ops/sec ±1.58% (87 runs sampled)
Fastest is v2#serialize
```
</details>

Resolves: https://github.com/prismicio/prismic-client/issues/198

## Checklist:

<!--- Put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [ ] My change requires an update to the official documentation.
- [x] All [TSDoc](https://tsdoc.org) comments are up-to-date and new ones have been added where necessary.
- [x] All new and existing tests are passing.

<!--- A cute animal (or car) picture is welcome to close your PR! -->
🐦