@angeloashmore 2021-11-29
Haha, you looked at it before I even had a chance to comment.

I think the `.sort()` method works better since it keeps the algorithm simpler. It's a bit slower like your benchmarks show, but not significantly.

We gain a tiny bit of performance back from [this change](https://github.com/prismicio/prismic-richtext/pull/51/files#diff-31305a1fcefd357bce204af67cdd145a94d63bf37cb7d6e18a62017e96df23edR151).